+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased](https://github.com/XaviArnaus/pyxavi/)
+
+### Added
+
+- New `Queue` module ([#29](https://github.com/XaviArnaus/pyxavi/pull/29))
+
 ## [v0.7.6](https://github.com/XaviArnaus/pyxavi/releases/tag/v0.7.6) - 2023-11-10
 
 ### Fixed